requirements.txt
Django==5.0.1
djangorestframework==3.14.0
langchain==0.3.12
langchain-openai==0.2.11
langgraph==0.2.60
python-dotenv==1.0.1
tiktoken==0.8.0
